Agent/requirements.txt
# LiveKit Agents Framework
livekit-agents==1.3.10
livekit-plugins-deepgram==1.3.10
livekit-plugins-openai==1.3.10

livekit-plugins-silero==1.3.10
livekit-plugins-simli~=1.3
livekit-plugins-turn-detector==1.3.10
livekit-plugins-noise-cancellation==0.2.5

# LlamaIndex for Knowledge Base
llama-index==0.14.10

# KB Pipeline - Document Processing
pymupdf>=1.24.0
python-docx>=1.1.0
openpyxl>=3.1.0
rank-bm25>=0.2.2
numpy>=1.26.0

# Utilities
python-dotenv==1.2.1
requests==2.32.5
pyyaml==6.0.2

# HTTP Clients
httpx==0.28.1
aiohttp==3.13.2

